remove old code, start unit test refactoring
