+++ src/cwe_checker_lib/src/intermediate_representation/expression.rs
@@ -1,9 +1,8 @@
-use std::collections::HashMap;
 use std::fmt::{self, Debug};
 
+use super::ByteSize;
 use super::Variable;
-use super::{ByteSize, Def};
-use crate::{pcode::RegisterProperties, prelude::*};
+use crate::prelude::*;
 
 mod builder;
 mod trivial_operation_substitution;
@@ -206,227 +205,6 @@ impl Expression {
             BinOp { lhs, rhs, .. } => {
                 lhs.substitute_input_var(input_var, replace_with_expression);
                 rhs.substitute_input_var(input_var, replace_with_expression);
-            }
-        }
-    }
-
-    /// This function checks for sub registers in pcode instruction and casts them into
-    /// SUBPIECE expressions with the base register as argument. It also checks whether
-    /// the given Term<Def> has a output sub register and if so, casts it into its
-    /// corresponding base register.
-    /// Lastly, it checks whether the following pcode instruction is a zero extension of
-    /// the currently overwritten sub register. If so, the zero extension is wrapped around
-    /// the current instruction and the TID of the zero extension instruction is returned
-    /// for later removal.
-    /// If there is no zero extension but an output register, the multiple SUBPIECEs are put
-    /// together to the size of the corresponding output base register using the PIECE instruction.
-    /// A few examples:
-    /// 1. From: EAX = COPY EDX;
-    ///    To:   RAX = COPY PIECE(SUBPIECE(RAX, 4, 4), SUBPIECE(RDX, 0, 4));
-    ///
-    /// 2. From:  AH = AH INT_XOR AH;
-    ///    To:   RAX = PIECE(PIECE(SUBPIECE(RAX, 2, 6), (SUBPIECE(RAX, 1, 1) INT_XOR SUBPIECE(RAX, 1, 1)), SUBPIECE(RAX, 0, 1));
-    ///
-    /// 3. FROM EAX = COPY EDX && RAX = INT_ZEXT EAX;
-    ///    To:  RAX = INT_ZEXT SUBPIECE(RDX, 0, 4);
-    pub fn cast_sub_registers_to_base_register_subpieces(
-        &mut self,
-        output: Option<&mut Variable>,
-        register_map: &HashMap<&String, &RegisterProperties>,
-        peeked: Option<&&mut Term<Def>>,
-    ) -> Option<Tid> {
-        let mut output_base_size: Option<ByteSize> = None;
-        let mut output_base_register: Option<&&RegisterProperties> = None;
-        let mut output_sub_register: Option<&RegisterProperties> = None;
-        let mut zero_extend_tid: Option<Tid> = None;
-
-        if let Some(output_value) = output {
-            if let Some(register) = register_map.get(&output_value.name) {
-                if *register.register != *register.base_register {
-                    output_sub_register = Some(register);
-                    output_base_register = register_map.get(&register.base_register);
-                    output_value.name = register.base_register.clone();
-                    output_value.size = output_base_register.unwrap().size;
-                    output_base_size = Some(output_value.size);
-
-                    if let Some(peek) = peeked {
-                        zero_extend_tid = peek.check_for_zero_extension(
-                            output_value.name.clone(),
-                            output_sub_register.unwrap().register.clone(),
-                        );
-                    }
-                }
-            }
-        }
-        self.replace_input_sub_register(register_map);
-        // based on the zero extension and base register output, either piece the subpieces together,
-        // zero extend the expression or do nothing (e.g. if output is a virtual register, no further actions should be taken)
-        self.piece_zero_extend_or_none(
-            zero_extend_tid.clone(),
-            output_base_register,
-            output_base_size,
-            output_sub_register,
-        );
-
-        zero_extend_tid
-    }
-
-    /// This function recursively iterates into the expression and checks whether a sub register was used.
-    /// If so, the sub register is turned into a SUBPIECE of the corresponding base register.
-    pub fn replace_input_sub_register(
-        &mut self,
-        register_map: &HashMap<&String, &RegisterProperties>,
-    ) {
-        match self {
-            Expression::BinOp { lhs, rhs, .. } => {
-                lhs.replace_input_sub_register(register_map);
-                rhs.replace_input_sub_register(register_map);
-            }
-            Expression::UnOp { arg, .. } | Expression::Cast { arg, .. } => {
-                arg.replace_input_sub_register(register_map)
-            }
-            Expression::Subpiece { arg, .. } => {
-                let truncated: &mut Expression = arg;
-                // Check whether the truncated data source is a sub register and if so,
-                // change it to its corresponding base register.
-                match truncated {
-                    Expression::Var(variable) => {
-                        if let Some(register) = register_map.get(&variable.name) {
-                            if variable.name != *register.base_register {
-                                variable.name = register.base_register.clone();
-                                variable.size =
-                                    register_map.get(&register.base_register).unwrap().size
-                            }
-                        }
-                    }
-                    _ => arg.replace_input_sub_register(register_map),
-                }
-            }
-            Expression::Var(variable) => {
-                if let Some(register) = register_map.get(&variable.name) {
-                    // We replace the register with a subpiece if the register itself is not a base register
-                    // or if the expression is an implicit subpiece (identifiable with `variable.size < register.size`).
-                    if variable.name != *register.base_register || variable.size < register.size {
-                        let target_size = variable.size;
-                        self.create_subpiece_from_sub_register(
-                            register.base_register.clone(),
-                            target_size,
-                            register.lsb,
-                            register_map,
-                        );
-                    }
-                }
-            }
-            _ => (),
-        }
-    }
-
-    /// This function creates a SUBPIECE expression
-    /// from a sub_register containing the corresponding base register.
-    fn create_subpiece_from_sub_register(
-        &mut self,
-        base: String,
-        size: ByteSize,
-        lsb: ByteSize,
-        register_map: &HashMap<&String, &RegisterProperties>,
-    ) {
-        *self = Expression::Subpiece {
-            low_byte: lsb,
-            size,
-            arg: Box::new(Expression::Var(Variable {
-                name: base.clone(),
-                size: register_map.get(&base).unwrap().size,
-                is_temp: false,
-            })),
-        };
-    }
-
-    /// This function either wraps the current expression into a
-    /// 1. zero extension expression: if the next instruction is a zero extension
-    /// of the currently overwritten sub register
-    /// 2. piece expression: if no zero extension is done the a sub register is overwritten
-    /// or does nothing in case there is no overwritten sub register.
-    fn piece_zero_extend_or_none(
-        &mut self,
-        zero_extend: Option<Tid>,
-        output_base_register: Option<&&RegisterProperties>,
-        output_size: Option<ByteSize>,
-        sub_register: Option<&RegisterProperties>,
-    ) {
-        if zero_extend.is_some() {
-            *self = Expression::Cast {
-                op: CastOpType::IntZExt,
-                size: output_size.unwrap(),
-                arg: Box::new(self.clone()),
-            }
-        } else if output_base_register.is_some() {
-            self.piece_two_expressions_together(
-                *output_base_register.unwrap(),
-                sub_register.unwrap(),
-            );
-        }
-    }
-
-    /// This function puts multiple SUBPIECE into PIECE of the size of the
-    /// base register. Depending on the position of the LSB of the sub register,
-    /// also nested PIECE instruction are possible.
-    fn piece_two_expressions_together(
-        &mut self,
-        output_base_register: &RegisterProperties,
-        sub_register: &RegisterProperties,
-    ) {
-        let base_size: ByteSize = output_base_register.size;
-        let base_name: &String = &output_base_register.register;
-        let sub_size: ByteSize = sub_register.size;
-        let sub_lsb: ByteSize = sub_register.lsb;
-
-        let base_subpiece = Box::new(Expression::Var(Variable {
-            name: base_name.clone(),
-            size: base_size,
-            is_temp: false,
-        }));
-
-        if sub_register.lsb > ByteSize::new(0) && sub_register.lsb + sub_register.size == base_size
-        {
-            // Build PIECE as PIECE(lhs: sub_register, rhs: low subpiece)
-            *self = Expression::BinOp {
-                op: BinOpType::Piece,
-                lhs: Box::new(self.clone()),
-                rhs: Box::new(Expression::Subpiece {
-                    low_byte: ByteSize::new(0),
-                    size: sub_lsb,
-                    arg: base_subpiece,
-                }),
-            }
-        } else if sub_register.lsb > ByteSize::new(0) {
-            // Build PIECE as PIECE(lhs:PIECE(lhs:higher subpiece, rhs:sub register), rhs:lower subpiece)
-            *self = Expression::BinOp {
-                op: BinOpType::Piece,
-                lhs: Box::new(Expression::BinOp {
-                    op: BinOpType::Piece,
-                    lhs: Box::new(Expression::Subpiece {
-                        low_byte: sub_lsb + sub_size,
-                        size: base_size - (sub_lsb + sub_size),
-                        arg: base_subpiece.clone(),
-                    }),
-                    rhs: Box::new(self.clone()),
-                }),
-                rhs: Box::new(Expression::Subpiece {
-                    low_byte: ByteSize::new(0),
-                    size: sub_lsb,
-                    arg: base_subpiece,
-                }),
-            }
-        } else {
-            // Build PIECE as PIECE(lhs: high subpiece, rhs: sub register)
-            *self = Expression::BinOp {
-                op: BinOpType::Piece,
-                lhs: Box::new(Expression::Subpiece {
-                    low_byte: sub_size,
-                    size: base_size - sub_size,
-                    arg: base_subpiece,
-                }),
-                rhs: Box::new(self.clone()),
             }
         }
     }
